refactor(compose): MailAddressInfo is now in its own file
